Initial work on 1722.1 ACQUIRE_ENTITY capability.
The control flow is messy and could be improved.

+++ module_avb_1722_1/src/avb_1722_1_common.c
@@ -5,7 +5,7 @@ extern unsigned char my_mac_addr[6];
 
 void set_guid(guid_t *guid, unsigned char *a)
 {
-	guid->l = (unsigned long long)a[0] + ((unsigned long long)a[1] << 32);
+	guid->l = (int)a[0] + ((int)a[1] << 32);
 }
 
 unsigned compare_guid(unsigned char *a, guid_t *b)
